refactor: TotalOrderedLottoGameNubmers 리팩토링
로또 구매 갯수를 반환하던 메서드를 order type에 따라 몇장 구매했는지 반환하는 메서드로 변경
OrderedLottoGameNumbers의 이름이 OrderedLottoGameDescription 변함에 따라
TotalOrderedLottoGameNubmers도 TotalOrderedLottoGameNubmers로 이름 변경
TestUtil에 manualLotto도 생성할 수 있는 메서드 추가
